# src/swagger/swagger-latest.yaml
swagger: '2.0'
info:
  contact: {}
paths:
...
  /api/v2/sign-in:
    post:
      description: sing in
      consumes:
        - application/json
      tags:
        - auth/v2
      summary: Sign in
      parameters:
        - description: User body
          name: input
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.SignInRequest'
        - type: string
          description: Admin auth secret
          name: admin_secret
          in: query
      responses:
        '200':
          description: Token
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.TokenResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
  /api/v2/sign-out:
    post:
      description: sing out
      consumes:
        - application/json
      tags:
        - auth/v2
      summary: Sign out
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
  /api/v2/sign-up:
    post:
      description: sing up
      consumes:
        - application/json
      tags:
        - auth/v2
      summary: Sign up
      parameters:
        - description: User body
          name: input
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.SignUpRequest'
        - type: string
          description: Admin auth secret
          name: admin_secret
          in: query
      responses:
        '201':
          description: ID
          schema:
            type: integer
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
...
definitions:
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: error description
...
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.SignInRequest:
    type: object
    properties:
      login:
        type: string
      password:
        type: string
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.SignUpRequest:
    type: object
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.TokenResponse:
    type: object
    properties:
      token:
        type: string
        example: jwt-format token